修改了文本json输入，添加了field的实现
